Implement the improved weight initialization
Initialize input weights of a neuron as random numbers with mean 0 and
standard deviation 1/sqrt(N), where N is the number of input weights.

Also add a weight and bias initializer trait and use it to implement two
initializers (the improved one and the naive one) so that we can compare
them conveniently.